Fixed typo, translated option, internal consistency
Changed Configuración to Ajustes as that is how it appears in the app. Changed a response in the FAQs to address the reader in usted form because that is how the other questions are addressed. Capitalized Galería Simple so it is a proper noun. Fixed spelling of completa. Translated to Spanish the option allow_deep_zooming_images. Changed to infinitive form the verb Mostrar in option show_recycle_bin.