Let ExoPlayer handle video looping seamlessly
Now that videos are replacing the inefficient GIF format, gapless loops are
important. ExoPlayer's built-in mechanism prebuffers the video to enable this,
whereas the current implementation of seeking to the start presents a short but
noticeable delay between each loop.

Note that this change introduces an incompatibility with current behaviour: due
to google/ExoPlayer#6459, certain videos with broken audio tracks that played
fine before will no longer start. Disabling the audio track is a workaround to
re-enable looping playback, but ExoPlayer does not appear to expose a way to
check if the audio track is short enough to produce the bug.